Further shrinking of SpookDog.choose_move(). WARNING: this slightly changes the AI's functionality from the original (I think it makes more sense though, and it makes the code simpler)
