+++ lib/spook_dog/spook_dog.py
@@ -31,17 +31,6 @@ class SpookDog(Opponent):
             moves = [move for move in moves if 0 <= self.current_mana - move.mana_cost <= self.max_mana]
             return moves
 
-        def random_weighted(moves_with_weights):
-            r = random.random()
-            upper_limit = 0
-            weight_sum = sum(moves_with_weights.values())
-            move = None
-            for move, weight in moves_with_weights.items():
-                upper_limit += weight/weight_sum
-                if r < upper_limit:
-                    return move
-            return move     # Off chance that weights don't sum perfectly to 1 due to computer rounding errors.
-
         if self.current_mana < 10:  # Only usable move
             return self.MOVE_TELEPORT
 
@@ -52,31 +41,23 @@ class SpookDog(Opponent):
         if self.game.player.current_hp <= 30:
             if self.current_mana < 50:
                 options = {self.MOVE_TELEPORT: 3, self.MOVE_GLIDE: 6}
-                if self.current_hp <= 180:
-                    options[self.MOVE_HEAL] = 1
-                return random_weighted(options)
-
-            if self.current_mana <= 140:
-                if self.current_hp <= 180:
-                    return random_weighted({self.MOVE_HEAL: 0.05, self.MOVE_TELEPORT: 0.1,
-                                            self.MOVE_GLIDE: 0.2, self.MOVE_CLAW: 0.65})
-                return random_weighted({self.MOVE_TELEPORT: 0.2, self.MOVE_GLIDE: 0.2, self.MOVE_CLAW: 0.6})
-
+            elif self.current_mana <= 140:
+                options = {self.MOVE_TELEPORT: 1, self.MOVE_GLIDE: 2, self.MOVE_CLAW: 6}
+            else:
+                options = {self.MOVE_GLIDE: 2.3333333333, self.MOVE_CLAW: 6.6666666667}
             if self.current_hp <= 180:
-                return random_weighted({self.MOVE_HEAL: 1/15, self.MOVE_GLIDE: 4/15, self.MOVE_CLAW: 10/15})
-            return random_weighted({self.MOVE_GLIDE: 0.25, self.MOVE_CLAW: 0.75})
+                options[self.MOVE_HEAL] = 1
+            return self.random_weighted(options)
 
         if self.current_hp < 25:
             if self.current_mana < 50:
-                return random_weighted({self.MOVE_TELEPORT: 0.1, self.MOVE_GLIDE: 0.1, self.MOVE_HEAL: 0.8})
-            if self.current_mana <= 140:
-                if self.game.player.current_hp <= 40:
-                    return random.choice([self.MOVE_HEAL, self.MOVE_CLAW])
-                return random_weighted({self.MOVE_HEAL: 0.85, self.MOVE_TELEPORT: 0.05,
-                                        self.MOVE_GLIDE: 0.05, self.MOVE_CLAW: 0.05})
+                return self.random_weighted({self.MOVE_TELEPORT: 0.1, self.MOVE_GLIDE: 0.1, self.MOVE_HEAL: 0.8})
             if self.game.player.current_hp <= 40:
                 return random.choice([self.MOVE_CLAW, self.MOVE_HEAL])
-            return random_weighted({self.MOVE_HEAL: 0.9, self.MOVE_GLIDE: 0.05, self.MOVE_CLAW: 0.05})
+
+            if random.random() < 0.1:
+                return random.choice(attack_options())
+            return self.MOVE_HEAL
 
         options = attack_options()
         if self.current_hp <= 0.9*self.max_hp:
